Correct onstart event order for MediaRecorder of mp4 muxer.
The onstart event of MediaRecorder should be triggered when
data is first written by the addition of a frame. However, the current
code only flushes the data during a timeout or at the end of recording,
causing the onstart event to be fired late.

To fix this issue, the static data of the ‘ftyp’ box should be written
as soon as the incoming frame is written to the internal cache.

Bug: 1520136
Change-Id: Ief9d67b7759fc50a18ccc3638f6825a8d47f71ee