[A11y] Do not include hidden nodes when they are ignored
The only time we need hidden nodes is when they are part of a
label or description. This CL keeps track of whether a given node
has ever been part of a label or description. If the node is hidden,
it removes it from the tree unless it's needed for a label/description.

In combination with CL:4852390, which removes hidden line breaks,
reduces the time spend in deserialization by an average of 32%
in perf tests.

Also improves:
- What's exposed for display:contents elements
- Support for aria-hidden="true" on iframes

Bug: 1448706, 1353788
Cq-Include-Trybots: luci.chromium.try:fuchsia-x64-accessibility-rel,linux-blink-web-tests-force-accessibility-rel
Change-Id: I1b0de1d9b6f037c107efd3ca93c073382fd0ece1