Router evaluation for subresources in the no fetch handler case
Before this CL, in the renderer, registered routers are not evaluated if
there is no fetch handler in the service worker. This is critical and
introduces unexpected behavior for the static routing API, because
developer may only use static routing API, that doesn't guarantee the
fetch handler existence.

This CL add the new boolean field, `need_router_evaluate` to
ControllerServiceWorkerInfo mojo interface. The renderer side use this
field to decide whether it involves ServiceWorkerSubresourceLoader or
not.

This CL also updates
`ServiceWorkerContainerHost::GetRemoteControllerServiceWorker()` not to
return nullptr when the router evaluation is needed, even if the
controller service worker does not have a fetch handler in the browser
side.

Bug: 1511459
Change-Id: I88e09900931a149c78cc4f25015fa5a5b1ffd62c