Rework text-shadow painting for text decorations
The current painting code-path for text-decorations with text-shadows is
complicated: line-throughs are painted using a draw-looper while the
rest of the decorations are painted for each shadow "layer" using a
DropShadowPaintFilter.

Unify these two by introducing a ScopedTextShadowPainter object that
constructs a filter - either a single DropShadowPaintFilter or a merge
filter combining one DropShadowPaintFilter for each layer. This makes
the two cases consistent, and for non-line-through decorations there's
no longer a need to run the paint code for each shadow "layer".

Bug: 1521809
Change-Id: I496d28672719e5efeee660dc780a96e433578456