[SAA] Implementing SharedWorker support (Part 2: Expose same-site cookie IDL)
The Storage Access API provides access to unpartitioned cookies in
third-party contexts. This CL is part of a series to extend that access
to SharedWorkers.

Before we actually provide access to SharedWorkers, we need a way for
the renderer to limit the scope of SameSite cookies provided to the
worker. Normally, a worker with a first-party storage key could access
all cookies, but if access is via the SAA handle we want to limit that
access to just SameSite: None cookies.

This CL is part of a series to allow that limiting ability. For now, we
expose the option in IDL but do not implement the change needed to limit
cookie access in first-party contexts when the option is set to 'none'.
The next CL will implement this limit.

Note: The ability to request 'all' in a third-party context is not added
here nor will it be added. We verify that an exception is thrown if this
is attempted.

Explainer:
https://privacycg.github.io/saa-non-cookie-storage/shared-workers.html

Part 1: Expose same-site cookie IPC
Part 2: Expose same-site cookie IDL
Part 3: Implement same-site cookie filtering
Part 4: Add SharedWorker support to SAA handle

Bug: 1484966
Change-Id: I8c56c02025950e119e3967afd49ed684446fd6e3